+++ ChangeLog.rst
@@ -18,7 +18,7 @@ Unreleased_
 
 Added
 -----
-- introduce ``pm-pattern-file-fmt`` configuration parameter to give full
+- Introduce ``pm-pattern-file-fmt`` configuration parameter to give full
   control over the path to pattern files.
 
 Fixed