+++ ChangeLog.rst
@@ -16,12 +16,18 @@ to `Semantic Versioning`_.
 Unreleased_
 ===========
 
+Added
+-----
+- Introduce ``pm-pattern-file-fmt`` configuration parameter to give full
+  control over the path to pattern files.
+
 Fixed
 -----
 
 - **BREAKING CHANGE** The expectation files path has never used the
   ``<test-module-name>`` component despite the ``README.rst`` claimed.
-
+  Existed projects could set ``pm-pattern-file-fmt`` to
+  ``{class}/{fn}{callspec}{suffix}`` to preserve backward compatibility.
 
 
 1.6.0_ -- 2024-02-29