+++ ChangeLog.rst
@@ -1,11 +1,11 @@
 .. SPDX-FileCopyrightText: 2017-now, See ``CONTRIBUTORS.lst``
 .. SPDX-License-Identifier: CC0-1.0
 
-.. program:: pytest-matcher
-
 ==========
 Change Log
 ==========
+
+.. option:: pm-mismatch-style
 
 All notable changes to this project will be documented in this file.